scripts/requirements.txt
google-generativeai
elevenlabs
requests
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
loguru
pydantic
